refactor(attention): the health lanes are their own file
lanes.go crossed the 500-line cap. The seams that report what WENT WRONG share
a shape the work lanes do not — optional, so an unbound one is absent rather
than empty, and per-reader, so a producer that cannot name a human is withheld
— and renderhealthlanes.go already draws them alone.
